refactor: no spartan; Object.fromEntries
