+++ stories/goal/5_minimal.tsx
