+++ stories/goal/6_minimal_horizontal.tsx
